BackgrundRemover/requirements.txt
pathlib==1.0.1
pillow==10.3.0
regex==2023.12.25
rembg==2.0.56